move lookup file
