+++ tests/sanity/ignore-2.10.txt
@@ -41,27 +41,18 @@ plugins/modules/turbo_fail.py import-2.6!skip
 plugins/modules/turbo_fail.py import-2.7!skip
 plugins/modules/turbo_fail.py import-3.5!skip
 plugins/modules/turbo_fail.py metaclass-boilerplate!skip
-plugins/module_utils/turbo/lookup.py compile-2.6!skip
-plugins/module_utils/turbo/lookup.py compile-2.7!skip
-plugins/module_utils/turbo/lookup.py compile-3.5!skip
-plugins/module_utils/turbo/lookup.py compile-3.6!skip
-plugins/module_utils/turbo/lookup.py compile-3.7!skip
-plugins/module_utils/turbo/lookup.py compile-3.8!skip
-plugins/module_utils/turbo/lookup.py future-import-boilerplate!skip
-plugins/module_utils/turbo/lookup.py import-2.6!skip
-plugins/module_utils/turbo/lookup.py import-2.7!skip
-plugins/module_utils/turbo/lookup.py import-3.5!skip
-plugins/module_utils/turbo/lookup.py import-3.6!skip
-plugins/module_utils/turbo/lookup.py import-3.7!skip
-plugins/module_utils/turbo/lookup.py import-3.8!skip
-plugins/module_utils/turbo/lookup.py import-3.9!skip
-plugins/module_utils/turbo/lookup.py metaclass-boilerplate!skip
-plugins/module_utils/turbo/lookup.py pylint:ansible-bad-module-import
+plugins/plugin_utils/turbo/lookup.py compile-2.6!skip
+plugins/plugin_utils/turbo/lookup.py compile-2.7!skip
+plugins/plugin_utils/turbo/lookup.py compile-3.5!skip
+plugins/plugin_utils/turbo/lookup.py compile-3.6!skip
+plugins/plugin_utils/turbo/lookup.py compile-3.7!skip
+plugins/plugin_utils/turbo/lookup.py compile-3.8!skip
+plugins/plugin_utils/turbo/lookup.py future-import-boilerplate!skip
+plugins/plugin_utils/turbo/lookup.py metaclass-boilerplate!skip
 plugins/lookup/turbo_demo.py compile-2.6!skip
 plugins/lookup/turbo_demo.py compile-2.7!skip
 plugins/lookup/turbo_demo.py compile-3.5!skip
 plugins/lookup/turbo_demo.py future-import-boilerplate!skip
-plugins/lookup/turbo_demo.py metaclass-boilerplate!skip
 plugins/module_utils/turbo/common.py compile-2.6!skip
 plugins/module_utils/turbo/common.py compile-2.7!skip
 plugins/module_utils/turbo/common.py compile-3.5!skip
